react redux rename component and container
