+++ client/src/components/stockBoard/stockTable/index.js
