+++ client/src/containers/tableHeader.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux'
 import { sortStocksData } from '../actions/stocksData'
-import Header from '../components/stockBoard/header'
+import TableHeader from '../components/stockBoard/tableHeader'
 
 const mapStateToProps = state => ({
   options: state.stockOption
@@ -10,9 +10,7 @@ const mapDispatchToProps = dispatch => ({
   onClick: i => dispatch(sortStocksData(i))
 })
 
-const StockTable = connect(
+export default connect(
   mapStateToProps,
   mapDispatchToProps
-)(Header)
-
-export default StockTable
+)(TableHeader)
